EchoClientSkeleton update
Managed to decrypt server's random bytes. Netx roadblock: Verifying
server's signature. How to do this if the client doesn't have access to
the server's ".pem" files to use PemUtils and read the server's public
signature key. We received the server's key text, but how to create a
public key for the cipher with that text?